Refactor package and party allocation
