+++ ledger/participant-state/kvutils/src/main/scala/com/daml/ledger/participant/state/kvutils/KeyValueCommitting.scala
@@ -5,7 +5,11 @@ package com.daml.ledger.participant.state.kvutils
 
 import com.daml.ledger.participant.state.kvutils.Conversions._
 import com.daml.ledger.participant.state.kvutils.DamlKvutils._
-import com.daml.ledger.participant.state.kvutils.committing.ProcessTransactionSubmission
+import com.daml.ledger.participant.state.kvutils.committing.{
+  ProcessPackageUpload,
+  ProcessPartyAllocation,
+  ProcessTransactionSubmission
+}
 import com.daml.ledger.participant.state.v1.Configuration
 import com.digitalasset.daml.lf.data.Ref.PackageId
 import com.digitalasset.daml.lf.data.Time.Timestamp
@@ -13,9 +17,6 @@ import com.digitalasset.daml.lf.engine.Engine
 import com.google.common.io.BaseEncoding
 import com.google.protobuf.ByteString
 import org.slf4j.LoggerFactory
-
-import scala.collection.JavaConverters._
-import scala.collection.breakOut
 
 object KeyValueCommitting {
   private val logger = LoggerFactory.getLogger(this.getClass)
@@ -91,20 +92,20 @@ object KeyValueCommitting {
     // Look at what kind of submission this is...
     submission.getPayloadCase match {
       case DamlSubmission.PayloadCase.PACKAGE_UPLOAD_ENTRY =>
-        processPackageUpload(
+        ProcessPackageUpload(
           entryId,
           recordTime,
           submission.getPackageUploadEntry,
           inputState
-        )
+        ).run
 
       case DamlSubmission.PayloadCase.PARTY_ALLOCATION_ENTRY =>
-        processPartyAllocation(
+        ProcessPartyAllocation(
           entryId,
           recordTime,
           submission.getPartyAllocationEntry,
           inputState
-        )
+        ).run
 
       case DamlSubmission.PayloadCase.CONFIGURATION_ENTRY =>
         logger.trace(
@@ -126,173 +127,11 @@ object KeyValueCommitting {
           submission.getTransactionEntry,
           inputLogEntries,
           inputState
-        ).result
+        ).run
 
       case DamlSubmission.PayloadCase.PAYLOAD_NOT_SET =>
         throw Err.InvalidPayload("DamlSubmission.payload not set.")
     }
-  }
-
-  private def processPackageUpload(
-      entryId: DamlLogEntryId,
-      recordTime: Timestamp,
-      packageUploadEntry: DamlPackageUploadEntry,
-      inputState: Map[DamlStateKey, Option[DamlStateValue]]
-  ): (DamlLogEntry, Map[DamlStateKey, DamlStateValue]) = {
-    val submissionId = packageUploadEntry.getSubmissionId
-
-    val archives = packageUploadEntry.getArchivesList.asScala
-
-    def tracelog(msg: String): Unit =
-      logger.trace(
-        s"""processPackageUpload[entryId=${prettyEntryId(entryId)}, submId=$submissionId], packages=${archives
-          .map(_.getHash)
-          .mkString(",")}]: $msg""")
-
-    // TODO: Add more comprehensive validity test, in particular, take the transitive closure
-    // of all packages being uploaded and see if they compile
-    archives.foldLeft[(Boolean, String)]((true, ""))(
-      (acc, archive) =>
-        if (archive.getPayload.isEmpty) (false, acc._2 ++ s"empty package '${archive.getHash}';")
-        else acc) match {
-
-      case (false, error) =>
-        tracelog(s"Package upload failed, invalid package submitted")
-        buildPackageRejectionLogEntry(
-          recordTime,
-          packageUploadEntry,
-          _.setInvalidPackage(
-            DamlPackageUploadRejectionEntry.InvalidPackage.newBuilder
-              .setDetails(error)))
-      case (_, _) =>
-        val filteredArchives = archives
-          .filter(
-            archive =>
-              inputState(
-                DamlStateKey.newBuilder
-                  .setPackageId(archive.getHash)
-                  .build).isEmpty
-          )
-        tracelog(s"Packages committed")
-        (
-          DamlLogEntry.newBuilder
-            .setRecordTime(buildTimestamp(recordTime))
-            .setPackageUploadEntry(
-              DamlPackageUploadEntry.newBuilder
-                .setSubmissionId(submissionId)
-                .addAllArchives(filteredArchives.asJava)
-                .setSourceDescription(packageUploadEntry.getSourceDescription)
-                .setParticipantId(packageUploadEntry.getParticipantId)
-                .build
-            )
-            .build,
-          filteredArchives
-            .map(
-              archive =>
-                (
-                  DamlStateKey.newBuilder.setPackageId(archive.getHash).build,
-                  DamlStateValue.newBuilder.setArchive(archive).build
-              )
-            )(breakOut)
-        )
-    }
-  }
-
-  private def buildPackageRejectionLogEntry(
-      recordTime: Timestamp,
-      packageUploadEntry: DamlPackageUploadEntry,
-      addErrorDetails: DamlPackageUploadRejectionEntry.Builder => DamlPackageUploadRejectionEntry.Builder
-  ): (DamlLogEntry, Map[DamlStateKey, DamlStateValue]) = {
-    (
-      DamlLogEntry.newBuilder
-        .setRecordTime(buildTimestamp(recordTime))
-        .setPackageUploadRejectionEntry(
-          addErrorDetails(
-            DamlPackageUploadRejectionEntry.newBuilder
-              .setSubmissionId(packageUploadEntry.getSubmissionId)
-              .setParticipantId(packageUploadEntry.getParticipantId)
-          ).build)
-        .build,
-      Map.empty
-    )
-  }
-
-  private def processPartyAllocation(
-      entryId: DamlLogEntryId,
-      recordTime: Timestamp,
-      partyAllocationEntry: DamlPartyAllocationEntry,
-      inputState: Map[DamlStateKey, Option[DamlStateValue]]
-  ): (DamlLogEntry, Map[DamlStateKey, DamlStateValue]) = {
-    val submissionId = partyAllocationEntry.getSubmissionId
-    def tracelog(msg: String): Unit =
-      logger.trace(
-        s"processPartyAllocation[entryId=${prettyEntryId(entryId)}, submId=$submissionId]: $msg")
-
-    val party: String = partyAllocationEntry.getParty
-    // 1. Verify that the party isn't empty
-    val partyValidityResult = !party.isEmpty
-    // 2. Verify that this is not a duplicate party submission.
-    val dedupKey = DamlStateKey.newBuilder
-      .setParty(party)
-      .build
-    val dedupEntry = inputState(dedupKey)
-    val dedupResult = dedupEntry.isEmpty
-
-    (partyValidityResult, dedupResult) match {
-      case (false, _) =>
-        tracelog(s"Party: $party allocation failed, party string invalid.")
-        buildPartyRejectionLogEntry(
-          recordTime,
-          partyAllocationEntry, {
-            _.setInvalidName(
-              DamlPartyAllocationRejectionEntry.InvalidName.newBuilder
-                .setDetails(s"Party string '$party' invalid"))
-          }
-        )
-      case (true, false) =>
-        tracelog(s"Party: $party allocation failed, duplicate party.")
-        buildPartyRejectionLogEntry(recordTime, partyAllocationEntry, {
-          _.setAlreadyExists(
-            DamlPartyAllocationRejectionEntry.AlreadyExists.newBuilder.setDetails(""))
-        })
-      case (true, true) =>
-        val key =
-          DamlStateKey.newBuilder.setParty(party).build
-        tracelog(s"Party: $party allocation committed.")
-        (
-          DamlLogEntry.newBuilder
-            .setRecordTime(buildTimestamp(recordTime))
-            .setPartyAllocationEntry(partyAllocationEntry)
-            .build,
-          Map(
-            key -> DamlStateValue.newBuilder
-              .setParty(
-                DamlPartyAllocation.newBuilder
-                  .setParticipantId(partyAllocationEntry.getParticipantId)
-                  .build)
-              .build
-          )
-        )
-    }
-  }
-
-  private def buildPartyRejectionLogEntry(
-      recordTime: Timestamp,
-      partyAllocationEntry: DamlPartyAllocationEntry,
-      addErrorDetails: DamlPartyAllocationRejectionEntry.Builder => DamlPartyAllocationRejectionEntry.Builder
-  ): (DamlLogEntry, Map[DamlStateKey, DamlStateValue]) = {
-    (
-      DamlLogEntry.newBuilder
-        .setRecordTime(buildTimestamp(recordTime))
-        .setPartyAllocationRejectionEntry(
-          addErrorDetails(
-            DamlPartyAllocationRejectionEntry.newBuilder
-              .setSubmissionId(partyAllocationEntry.getSubmissionId)
-              .setParticipantId(partyAllocationEntry.getParticipantId)
-          ).build)
-        .build,
-      Map.empty
-    )
   }
 
 }
