--- ledger/participant-state/kvutils/src/main/scala/com/daml/ledger/participant/state/kvutils/committing/CommitPackage.scala
@@ -1,3 +1,0 @@
-package com.daml.ledger.participant.state.kvutils.committing
-
-class CommitPackage {}
