+++ ledger/participant-state/kvutils/src/main/scala/com/daml/ledger/participant/state/kvutils/committing/Common.scala
@@ -1,0 +1,16 @@
+package com.daml.ledger.participant.state.kvutils.committing
+
+import com.daml.ledger.participant.state.kvutils.DamlKvutils.{
+  DamlLogEntry,
+  DamlStateKey,
+  DamlStateValue
+}
+
+object Common {
+  type DamlStateMap = Map[DamlStateKey, DamlStateValue]
+
+  // A check result, which is either a rejection or passing check with associated new state.
+  type CheckResult = Either[DamlLogEntry, DamlStateMap]
+  def pass(state: (DamlStateKey, DamlStateValue)*): CheckResult = Right(state.toMap)
+
+}
